* [MOSIP-24254] reverted pom version

* Fixed MOSIP-22787

* Revert "[MOSIP-24254] reverted pom version"

* MOSIP-25677 / MOSIP-25687 (#532)

* Fixed MOSIP-25677 / MOSIP-25687

* MOSIP-25677 / MOSIP-25687: Rolledback DB changes and modified code

* Rolled back tested change

* Changed version to 1.2.0.1-B2

* Modified class file

* Modified class file

* Replaced test file

* Modified test class

* Modified pom versions


* Changed registration-test pom version (#533)

* Fixed MOSIP-25677 / MOSIP-25687

* MOSIP-25677 / MOSIP-25687: Rolledback DB changes and modified code

* Rolled back tested change

* Changed version to 1.2.0.1-B2

* Modified class file

* Modified class file

* Replaced test file

* Modified test class

* Modified pom versions

* Changed registration-test pom version


* MOSIP-25677 / MOSIP-25687 (#544)


* Modified biometrics-util version (#545)

* MOSIP-25677 / MOSIP-25687

* Modified biometrics-util version

---------


* MOSIP-27529, MOSIP-27684 (#564)

* MOSIP-27529: MOSIP-27809, MOSIP-27684 (#562)

* MOSIP-27809: Upgrade changes

* Fixed MOSIP-27684

* MOSIP-27809: Handling inputstream

* Modified logging

* Added error log in catch block

---------

# Conflicts:
#	registration/registration-services/src/test/java/io/mosip/registration/test/update/SoftwareUpdateHandlerTest.java

* Added default-version-mappings

---------


* MOSIP-27529: Fixed test case (#565)

* MOSIP-27529: MOSIP-27809, MOSIP-27684 (#562)

* MOSIP-27809: Upgrade changes

* Fixed MOSIP-27684

* MOSIP-27809: Handling inputstream

* Modified logging

* Added error log in catch block

---------

# Conflicts:
#	registration/registration-services/src/test/java/io/mosip/registration/test/update/SoftwareUpdateHandlerTest.java

* Added default-version-mappings

* Fixed test case

---------


* Build issue resolution (#568)

* [MOSIP-23218] Updated Pom.xml version. (#569)

* Changed commons-packet-manager version to 1.2.0.1-SNAPSHOT

* Biometrics Changes (#567)

* MOSIP-27529: MOSIP-27809, MOSIP-27684 (#562)

* MOSIP-27809: Upgrade changes

* Fixed MOSIP-27684

* MOSIP-27809: Handling inputstream

* Modified logging

* Added error log in catch block

---------

# Conflicts:
#	registration/registration-services/src/test/java/io/mosip/registration/test/update/SoftwareUpdateHandlerTest.java

* Added default-version-mappings

* Fixed test case

* Display MDS and SDK quality when sdk quality check is enabled

# Conflicts:
#	registration/registration-client/src/main/java/io/mosip/registration/controller/GenericController.java
#	registration/registration-client/src/main/java/io/mosip/registration/controller/device/GenericBiometricsController.java
#	registration/registration-client/src/main/resources/application.css
#	registration/registration-client/src/main/resources/fxml/BiometricsCapture.fxml
#	registration/registration-services/src/main/java/io/mosip/registration/dto/RegistrationDTO.java

* Handling exception photo scan with new MDS

* Biometrics screen alignment modifications

* Rollback BiometricsCapture.fxml

* Modified biometrics screen alignments

* Rolled back RegistrationDTO.java

* Changes to display MDS and SDK score

* Changes to display MDS and SDK score

* Rolledback GenericBiometricsController.java

* Changes to display MDS and SDK scores

---------


* MOSIP-28152: Updated versions to 1.2.0.1-SNAPSHOT

* MOSIP-28152: Added parent pom (#572)


* MOSIP-28152 (#573)


* MOSIP-28152 (#574)

* MOSIP-28152: Added parent pom

* MOSIP-28152

---------


* MOSIP-27529: Updated scripts (#579)

* MOSIP-27529: Updated scripts

* Updated rollback script

* Updated default version mappings

---------


* MOSIP-27529


* MOSIP-27529: Modified upgrade scripts (#581)

* MOSIP-27529: Updated scripts

* Updated rollback script

* Updated default version mappings

* MOSIP-27529: Modified upgrade scripts

* Modified conditions

---------


* Fixed NullPointerException (#582)

* MOSIP-27529: Updated scripts

* Updated rollback script

* Updated default version mappings

* MOSIP-27529: Modified upgrade scripts

* Modified conditions

* Fixed NullPointerException

* Rolledback change

---------


* MOSIP-27529: Corrected upgrade script (#583)

* MOSIP-27529: Updated scripts

* Updated rollback script

* Updated default version mappings

* MOSIP-27529: Modified upgrade scripts

* Modified conditions

* Fixed NullPointerException

* Rolledback change

* MOSIP-27529: Corrected upgrade script

* MOSIP-27529: Added missing script

---------


* Added backward compatibility for schema (#584)

* MOSIP-27529: Updated scripts

* Updated rollback script

* Updated default version mappings

* MOSIP-27529: Modified upgrade scripts

* Modified conditions

* Fixed NullPointerException

* Rolledback change

* MOSIP-27529: Corrected upgrade script

* MOSIP-27529: Added missing script

* Added backward compatibility for schema

---------


* Template data map, compatibility added for 1.1.4 & 1.1.5.5 templates

* MOSIP-27529: Added changes in run.bat to support upgrade (#586)

* MOSIP-27529: Updated scripts

* Updated rollback script

* Updated default version mappings

* MOSIP-27529: Modified upgrade scripts

* Modified conditions

* Fixed NullPointerException

* Rolledback change

* MOSIP-27529: Corrected upgrade script

* MOSIP-27529: Added missing script

* Added backward compatibility for schema

* Added changes in run.bat to support upgrade

* MOSIP-27529: Added changes in run.bat to support upgrade

---------


* MOSIP-28152 / MOSIP-27529 (#587)

* MOSIP-28152 / MOSIP-27529: Removed default-version-mappings and updated
version-mappings value

* Updated property name

* Updated comments for version-mappings property

* Updated comments for version-mappings property

---------


* MOSIP-28152 / MOSIP-27529: Rolling back last change, added default-version-mappings (#588)

* MOSIP-28152 / MOSIP-27529: Removed default-version-mappings and updated
version-mappings value

* Updated property name

* Updated comments for version-mappings property

* Updated comments for version-mappings property

* MOSIP-28152 / MOSIP-27529: Rollingback last change, added
default-version-mappings

---------


* MOSIP-28705: Removed default-version-mappings and referring to
version-mappings in spring.properties

* Updated 1.1.5.5 released version in version-mappings

* Fixed MOSIP-28705 (#590)

* MOSIP-28705: Removed default-version-mappings and referring to
version-mappings in spring.properties

* Updated 1.1.5.5 released version in version-mappings

---------


* Updated comments

* Fixed test case

* Fixed MOSIP-28701

* MOSIP-28701: Added null check

* [MOSIP-29165] Updated workflows as per reusable workflow templates (#601)

[MOSIP-25631]


* [MOSIP-30718] Updated push-trigger.yml to fix docker build issue (#609)


* [MOSIP-30718] Updated push-trigger.yml to fix docker build issue (#610)


* [DSD-3932] removed release-changes.yml, tag.yml and updated push-trigger.yml file (#612)


* Release changes (#614)

* Updated Pom versions for release changes


* Update pom.xml


---------


* MOSIP-22171: Setting deviceType based on configuration


* MOSIP-29676: Deleting user biometrics on deactivation


* Reverted back to SNAPSHOT versions


* MOSIP-24220, MOSIP-25200: Sonar Fixes (#618)

* Fixed: MOSIP-24220

* Fixed: MOSIP-24220

* Fixed:MOSIP-24220

* Fixed: MOSIP-25200

* Fixed test case error


* Updated exception log

* Update MosipDeviceSpecification_095_ProviderImpl.java

---------


* MOSIP-31328: Rolled back sonar fix


* MOSIP-31900: Fixed build failure (#625)


* Updated Pom versions for release changes (#626)


* Resolved conflicts


* Fixed test case


---------

+++ registration/registration-services/src/main/java/io/mosip/registration/audit/AuditManagerSerivceImpl.java
@@ -137,7 +137,7 @@ public class AuditManagerSerivceImpl extends BaseService implements AuditManager
 
 		String val = getGlobalConfigValueOf(RegistrationConstants.AUDIT_TIMESTAMP);
 
-		if (val != null) {
+		if (val != null && !val.isEmpty()) {
 			try {
 				/* Delete Audits before given Time */
 				auditDAO.deleteAudits(DateUtils.parseToLocalDateTime(val));
